backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-multipart
sqlalchemy
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
python-dotenv
